chore: delete no longer needed username takeover screen
