+++ wire-ios-sync-engine/Source/Data Model/ZMUser+Consent.swift
@@ -62,13 +62,16 @@ extension ZMUser {
                       on transportSession: TransportSessionType,
                       completion: @escaping CompletionFetch) {
 
-        guard let apiVersion = BackendInfo.apiVersion else {
+        guard
+            let apiVersion = BackendInfo.apiVersion,
+            let context = managedObjectContext
+        else {
             return completion(.failure(ConsentRequestError.unknown))
         }
 
         let request = ConsentRequestFactory.fetchConsentRequest(apiVersion: apiVersion)
 
-        request.add(ZMCompletionHandler(on: managedObjectContext!) { response in
+        request.add(ZMCompletionHandler(on: context ) { response in
 
             guard 200 ... 299 ~= response.httpStatus,
                   let payload = response.payload
