+++ wire-ios/Wire-iOS/Sources/UserInterface/ConversationList/Container/ConversationListViewController+PushPermissions.swift
@@ -75,4 +75,8 @@ extension ConversationListViewController {
                                                object: nil)
 
     }
+
+    private func concealContentContainer() {
+        contentContainer.alpha = 0
+    }
 }
