+++ wire-ios/Wire-iOS/Sources/UserInterface/ConversationList/Container/ConversationListViewController.swift
@@ -44,7 +44,6 @@ final class ConversationListViewController: UIViewController {
     var startCallToken: Any?
 
     var pushPermissionDeniedViewController: PermissionDeniedViewController?
-    var usernameTakeoverViewController: UserNameTakeOverViewController?
 
     private let noConversationLabel: UILabel = {
         let label = UILabel()
@@ -153,7 +152,7 @@ final class ConversationListViewController: UIViewController {
         super.viewWillAppear(animated)
 
         viewModel.savePendingLastRead()
-        viewModel.requestSuggestedHandlesIfNeeded()
+        viewModel.requestMarketingConsentIfNeeded()
     }
 
     override func viewDidAppear(_ animated: Bool) {
@@ -365,10 +364,9 @@ final class ConversationListViewController: UIViewController {
                                      completion: completion)
     }
 
-    var hasUsernameTakeoverViewController: Bool {
-        return usernameTakeoverViewController != nil
+    func showNewsletterSubscriptionDialogIfNeeded(completionHandler: @escaping ResultHandler) {
+        UIAlertController.showNewsletterSubscriptionDialogIfNeeded(presentViewController: self, completionHandler: completionHandler)
     }
-
 }
 
 // MARK: - UITabBarDelegate
